backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.7.0
python-dotenv>=1.0.1
reportlab>=4.2.0
